chore: fix ios demo mode
