+++ detox/jest.detox.startup.js
@@ -49,8 +49,6 @@ beforeAll(async () => {
         execSync(`adb -s ${id} shell setprop debug.hwui.renderer skiag`);
         execSync(`adb -s ${id} reverse tcp:8080 tcp:8080`);
     }
-
-    await setDemoMode();
 }, 1800000);
 
 beforeEach(async () => {
@@ -61,30 +59,3 @@ afterAll(async () => {
     await adapter.afterAll();
     await cleanup();
 });
-
-async function setDemoMode() {
-    if (device.getPlatform() === "ios") {
-        const type = device.name.substring(device.name.indexOf("(") + 1, device.name.lastIndexOf(")"));
-        execSync(
-            `xcrun simctl status_bar "${type}" override --time "12:00" --batteryState charged --batteryLevel 100 --wifiBars 3 --cellularMode active --cellularBars 4`
-        );
-    } else {
-        const id = device.id;
-        // enter demo mode
-        execSync(`adb -s ${id} shell settings put global sysui_demo_allowed 1`);
-        // display time 12:00
-        execSync(`adb -s ${id} shell am broadcast -a com.android.systemui.demo -e command clock -e hhmm 1200`);
-        // Display full mobile data with 4g type and no wifi
-        execSync(
-            `adb -s ${id} shell am broadcast -a com.android.systemui.demo -e command network -e mobile show -e level 4 -e datatype 4g -e wifi false`
-        );
-        // Hide notifications
-        execSync(
-            `adb -s ${id} shell am broadcast -a com.android.systemui.demo -e command notifications -e visible false`
-        );
-        // Show full battery but not in charging state
-        execSync(
-            `adb -s ${id} shell am broadcast -a com.android.systemui.demo -e command battery -e plugged false -e level 100`
-        );
-    }
-}
